+++ turkey_address/__init__.py
@@ -1,2 +1,3 @@
-import res_partner
 
+import address_details
+import res_partner
