+++ turkey_address/res_partner_view.xml
@@ -1,89 +1,13 @@
 <?xml version="1.0"?>
 <openerp>
     <data>
-        <record id="view_address_district_tree" model="ir.ui.view">
-            <field name="name">view.address.region.tree</field>
-            <field name="model">address.district</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="name"/>
-                    <field name="state_id" />
-                </tree>
-            </field>
-        </record>
-        <record id="view_address_district_form" model="ir.ui.view">
-            <field name="name">view.address.district.form</field>
-            <field name="model">address.district</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <group col="4">
-                            <field name="name" required="1"/>
-                            <field name="state_id" required="1"/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-        <record id="view_address_region_tree" model="ir.ui.view">
-            <field name="name">view.address.region.tree</field>
-            <field name="model">address.region</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="name"/>
-                    <field name="district_id"/>
-                </tree>
-            </field>
-        </record>
-        <record id="view_address_region_form" model="ir.ui.view">
-            <field name="name">view.address.region.form</field>
-            <field name="model">address.region</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <group col="4">
-                            <field name="name" required="1"/>
-                            <field name="district_id" required="1"/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-        <record id="view_address_neighbour_tree" model="ir.ui.view">
-            <field name="name">view.address.neighbour.tree</field>
-            <field name="model">address.neighbour</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="name"/>
-                    <field name="region_id" />
-                    <field name="code" />
-                </tree>
-            </field>
-        </record>
-        <record id="view_address_neighbour_form" model="ir.ui.view">
-            <field name="name">view.address.neighbour.form</field>
-            <field name="model">address.neighbour</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <group col="4">
-                            <field name="name" required="1"/>
-                            <field name="region_id" required="1"/>
-                            <field name="code" required="1"/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
 
         <record id="view_res_partner_inherite_view" model="ir.ui.view">
             <field name="name">res.partner.inherit.view</field>
             <field name="model">res.partner</field>
             <field name="inherit_id" ref="base.view_partner_form"/>
             <field name="arch" type="xml">
-               <field name="country_id" position="after">
+                <field name="country_id" position="after">
                     <field name="district_id" placeholder="District"
                         on_change="onchange_district(district_id)"
                         context="{'state_id': state_id}"/>
@@ -91,63 +15,22 @@
                         on_change="onchange_region(region_id)"
                         context="{'district_id': district_id}"/>
                     <field name="neighbour_id" placeholder="Neighbour"
-                        context="{'region_id': region_id, 'zip': zip}"/>
-               </field>
+                        on_change="onchange_neighbour(neighbour_id)"
+                        context="{'region_id': region_id}"/>
+                </field>
+                <xpath expr="//form/sheet/notebook/page[1]/field/form/sheet/group[2]/div[2]/field[3]" position="after">
+                    <field name="district_id" placeholder="District"
+                        on_change="onchange_district(district_id)"
+                        context="{'state_id': state_id}"/>
+                    <field name="region_id" placeholder="Region"
+                        on_change="onchange_region(region_id)"
+                        context="{'district_id': district_id}"/>
+                    <field name="neighbour_id" placeholder="Neighbour"
+                        on_change="onchange_neighbour(neighbour_id)"
+                        context="{'region_id': region_id}"/>
+                </xpath>
             </field>
         </record>
 
-        <record id="action_district" model="ir.actions.act_window">
-            <field name="name">Districts</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">address.district</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[]</field>
-            <field name="help" type="html">
-              <p class="oe_view_nocontent_create">
-                Click to add a District in your Distriicts list.
-              </p>
-            </field>
-        </record>
-        <menuitem id="district_menu" action="action_district" parent="base.menu_localisation" sequence="5"/>
-
-        <record id="action_region" model="ir.actions.act_window">
-            <field name="name">Regions</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">address.region</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[]</field>
-            <field name="help" type="html">
-              <p class="oe_view_nocontent_create">
-                Click to add a Region in your Regions list.
-              </p>
-            </field>
-        </record>
-        <menuitem id="region_menu" action="action_region" parent="base.menu_localisation" sequence="6"/>
-
-        <record id="action_neighbour" model="ir.actions.act_window">
-            <field name="name">Neighbours</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">address.neighbour</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[]</field>
-            <field name="help" type="html">
-              <p class="oe_view_nocontent_create">
-                Click to add a Neighbour in your Neighbours list.
-              </p>
-            </field>
-        </record>
-        <menuitem id="neighbour_menu" action="action_neighbour" parent="base.menu_localisation" sequence="7"/>
-
-   </data>
+    </data>
 </openerp>
-
-
-
-
-
-
-
-
